Fix 'Edit Post' bug, which required us to make API respond with an ID so that the ID  can be injected into  and without the need for a page reload.

+++ client/app/scripts/services/postfactory.js
@@ -14,25 +14,39 @@ angular.module('clientApp')
     var postFactory= {};
 
     // Public API here
+    
     // CREATE
     postFactory.createPost = function (post) {
       console.log("it hit the factory...");
       return $http.post(urlBase, post);
     };
-    //READ
+    //READ - All posts
     postFactory.getPosts = function () {
       return $http.get(urlBase);
     };
+    //READ - Single post
     postFactory.getPost = function (id) {
-      return $http.get(urlBase + '/' + id);
+      if (id) {
+        return $http.get(urlBase + '/' + id);
+      } else {
+        console.error('No id given', data);
+      }
     };
     //UPDATE
     postFactory.updatePost = function (id, data) {
-      return $http.put(urlBase + '/' + id, data);
+      if (id){
+        return $http.put(urlBase + '/' + id, data);
+      } else {
+        console.error('No id given', data);
+      }
     };
     //DELETE
     postFactory.deletePost = function (id) {
-      return $http.delete(urlBase + '/' + id);
+      if (id) {
+        return $http.delete(urlBase + '/' + id);
+      } else {
+        console.error('No id given', data);
+      }
     };
 
     // return Factory to caller
